`StateManager` is completely intependent of the game now
This is the [same change][0] I did on `nsnake`.
The idea is to make the `Engine` stuff independent of the
game so we can reuse it between `nsnake`, `yetris` and every
other console game around.

The game crashes because I need to change how the GameStates
go from one another, it shall be on the next commit.

[0]: https://github.com/alexdantas/nSnake/commit/cc595d2f79bfa3668e0f56316437c02adcd22989