HUGE change, moved all Engine stuff to a different directory
Just like nsnake, the Game Engine is separated from the game.
This way we can unify a Game Engine made for creating console
games!

This is the first step, though, and the game quite doesn't
work well. Mainly the colors, but it should be okay after
I figure out what's the problem.

Anyways, now the code is pretty much the same between the
games! I should create another repository for this... Or
not!